+++ cmd/il/repro.go
@@ -7,8 +7,8 @@ import (
 	"path/filepath"
 	"strings"
 
-	"github.com/anomalyco/intent-lang/internal/agent"
-	"github.com/anomalyco/intent-lang/internal/il"
+	"github.com/lovepk/intent-lang/internal/agent"
+	"github.com/lovepk/intent-lang/internal/il"
 )
 
 func repro(args []string) error {